+++ grid2op/Parameters.py
@@ -106,6 +106,12 @@ class Parameters:
         Number of steps for which it's worth it to give an alarm (if an alarm is send outside of the window
         `[ALARM_BEST_TIME - ALARM_WINDOW_SIZE, ALARM_BEST_TIME + ALARM_WINDOW_SIZE]` then it does not grant anything
 
+    MAX_SIMULATE_PER_STEP: ``int``
+        Maximum number of calls to `obs.simuate(...)` allowed per step (reset each "env.step(...)"). Defaults to -1 meaning "as much as you want".
+
+    MAX_SIMULATE_PER_EPISODE: ``int``
+        Maximum number of calls to `obs.simuate(...)` allowed per episode (reset each "env.simulate(...)"). Defaults to -1 meaning "as much as you want".
+
     """
     def __init__(self, parameters_path=None):
         """
@@ -170,6 +176,9 @@ class Parameters:
 
         self.ALARM_BEST_TIME = 12
         self.ALARM_WINDOW_SIZE = 12
+
+        self.MAX_SIMULATE_PER_STEP = dt_int(-1)
+        self.MAX_SIMULATE_PER_EPISODE = dt_int(-1)
 
     @staticmethod
     def _isok_txt(arg):
@@ -272,6 +281,12 @@ class Parameters:
             warnings.warn("Parameters: The _parameters \"{}\" used to build the Grid2Op.Parameters "
                           "class are not recognized and will be ignored.".format(ignored_keys))
 
+        if "MAX_SIMULATE_PER_STEP" in dict_:
+            self.MAX_SIMULATE_PER_STEP = dt_int(dict_["MAX_SIMULATE_PER_STEP"])
+
+        if "MAX_SIMULATE_PER_EPISODE" in dict_:
+            self.MAX_SIMULATE_PER_EPISODE = dt_int(dict_["MAX_SIMULATE_PER_EPISODE"])
+
     def to_dict(self):
         """
         Serialize all the _parameters as a dictionnary; Useful to write it in json format.
@@ -299,6 +314,8 @@ class Parameters:
         res["ACTIVATE_STORAGE_LOSS"] = bool(self.ACTIVATE_STORAGE_LOSS)
         res["ALARM_BEST_TIME"] = int(self.ALARM_BEST_TIME)
         res["ALARM_WINDOW_SIZE"] = int(self.ALARM_WINDOW_SIZE)
+        res["MAX_SIMULATE_PER_STEP"] = int(self.MAX_SIMULATE_PER_STEP)
+        res["MAX_SIMULATE_PER_EPISODE"] = int(self.MAX_SIMULATE_PER_EPISODE)
         return res
 
     def init_from_json(self, json_path):
@@ -463,3 +480,19 @@ class Parameters:
             raise RuntimeError("self.ALARM_WINDOW_SIZE should be a positive integer !")
         if self.ALARM_BEST_TIME <= 0:
             raise RuntimeError("self.ALARM_BEST_TIME should be a positive integer !")
+
+        try:
+            self.MAX_SIMULATE_PER_STEP = int(self.MAX_SIMULATE_PER_STEP)  # to raise if numpy array
+            self.MAX_SIMULATE_PER_STEP = dt_int(self.MAX_SIMULATE_PER_STEP)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert MAX_SIMULATE_PER_STEP to int with error \n:\"{exc_}\"")
+        if self.MAX_SIMULATE_PER_STEP <= -2:
+            raise RuntimeError(f"self.MAX_SIMULATE_PER_STEP should be a positive integer or -1, we found {self.MAX_SIMULATE_PER_STEP}")
+
+        try:
+            self.MAX_SIMULATE_PER_EPISODE = int(self.MAX_SIMULATE_PER_EPISODE)  # to raise if numpy array
+            self.MAX_SIMULATE_PER_EPISODE = dt_int(self.MAX_SIMULATE_PER_EPISODE)
+        except Exception as exc_:
+            raise RuntimeError(f"Impossible to convert MAX_SIMULATE_PER_EPISODE to int with error \n:\"{exc_}\"")
+        if self.MAX_SIMULATE_PER_EPISODE <= -2:
+            raise RuntimeError(f"self.MAX_SIMULATE_PER_EPISODE should be a positive integer or -1, we found {self.MAX_SIMULATE_PER_EPISODE}")